+++ spin2/server/src/providers/SemanticTokensProvider.ts
@@ -63,7 +63,8 @@ export default class SemanticTokensProvider implements Provider {
     'pasmInline',
     'instance',
     'missingDeclaration',
-    'illegalUse'
+    'illegalUse',
+    'bitfield'
   ];
 
   private isDebugLogEnabled: boolean = false; // WARNING (REMOVE BEFORE FLIGHT)- change to 'false' - disable before commit